+++ main.cu
@@ -8,19 +8,27 @@
 
 
 int main() {
-    size_t m = 4;
-    size_t n = 3;
-    size_t k = 2;
-    std::vector<real_t> data{1, 2, 3,
-                             4, 5, 6,
-                             7, 8, 9,
-                             10, 11, 12,
-                             13, 14, 15,
-                             16, 17, 18,
-                             19, 20, 21,
-                             22, 23, 24};
-    DTensor<real_t> A(data, m, n, k, rowMajor);
-    std::cout << A << "\n";
+    size_t m = 3;
+    size_t n = 7;
+    std::vector<real_t> mat{1, -2, 3, 4, -1, -1, -1,
+                            1, 2, -3, 4, -1, -1, -1,
+                            -1, 3, 5, -7, -1, -1, -1};
+    DTensor mats(m, n, 1);
+    mats.upload(mat, rowMajor);
+
+    Nullspace ns = Nullspace(mats);
+
+    std::vector<real_t> vec{1, 2, 3, 4, 5, 6, 7};
+    DTensor vecs(vec, n);
+
+    ns.project(vecs);
+
+    std::cout << mats << "\n";
+    std::cout << vecs << "\n";
+
+    DTensor op(m, 1, 1);
+    op.addAB(mats, vecs);
+    std::cout << op << "\n";
 
     return 0;
 }